requirements.txt
web3
streamlit
pandas
plotly
networkx
python-dotenv
eth-account
py-solc-x
